got all crud to work with every case except session

+++ src/main/java/springboot/repository/PeopleRepository.java
@@ -3,5 +3,5 @@ package springboot.repository;
 import springboot.model.People;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PeopleRepository extends JpaRepository <People, Long>  {
+public interface PeopleRepository extends JpaRepository <People, Integer>  {
 }
